+++ src/services/qualidade.service.js
@@ -19,3 +19,18 @@ export const cadastraEmbalagens = async payload =>
 
 export const editaEmbalagem = async (payload, uuid) =>
   await axios.put(`/embalagens/${uuid}/`, payload);
+
+export const getUnidadesMedida = async params =>
+  await axios.get("/unidades-medida-logistica/", { params });
+
+export const getUnidadeMedida = async uuid =>
+  await axios.get(`/unidades-medida-logistica/${uuid}/`);
+
+export const cadastrarUnidadeMedida = async payload =>
+  await axios.post("/unidades-medida-logistica/", payload);
+
+export const editarUnidadeMedida = async (payload, uuid) =>
+  await axios.patch(`/unidades-medida-logistica/${uuid}/`, payload);
+
+export const getNomesEAbreviacoesUnidadesMedida = async () =>
+  await axios.get("/unidades-medida-logistica/lista-nomes-abreviacoes/");
